style: format glm4 modeling.py with black

+++ paddleformers/transformers/glm4/modeling.py
@@ -109,10 +109,7 @@ class Glm4RotaryEmbedding(nn.Layer):
         if self.rotary_dim % 2 != 0:
             self.rotary_dim -= 1
 
-        inv_freq = 1.0 / (
-            self.base
-            ** (paddle.arange(0, self.rotary_dim, 2, dtype="float32") / self.rotary_dim)
-        )
+        inv_freq = 1.0 / (self.base ** (paddle.arange(0, self.rotary_dim, 2, dtype="float32") / self.rotary_dim))
         self.register_buffer("inv_freq", inv_freq, persistable=False)
 
     def forward(self, position_ids):
@@ -193,15 +190,11 @@ class Glm4Attention(nn.Layer):
         key_states = self.k_proj(hidden_states)
         value_states = self.v_proj(hidden_states)
 
-        query_states = query_states.reshape(
-            [bsz, q_len, self.num_heads, self.head_dim]
-        ).transpose([0, 2, 1, 3])
-        key_states = key_states.reshape(
-            [bsz, q_len, self.num_key_value_heads, self.head_dim]
-        ).transpose([0, 2, 1, 3])
-        value_states = value_states.reshape(
-            [bsz, q_len, self.num_key_value_heads, self.head_dim]
-        ).transpose([0, 2, 1, 3])
+        query_states = query_states.reshape([bsz, q_len, self.num_heads, self.head_dim]).transpose([0, 2, 1, 3])
+        key_states = key_states.reshape([bsz, q_len, self.num_key_value_heads, self.head_dim]).transpose([0, 2, 1, 3])
+        value_states = value_states.reshape([bsz, q_len, self.num_key_value_heads, self.head_dim]).transpose(
+            [0, 2, 1, 3]
+        )
 
         if position_ids is None:
             position_ids = paddle.arange(q_len, dtype="int64").unsqueeze(0).tile([bsz, 1])
@@ -211,9 +204,7 @@ class Glm4Attention(nn.Layer):
         else:
             cos, sin = position_embeddings
 
-        query_states, key_states = apply_rotary_pos_emb(
-            query_states, key_states, cos, sin, unsqueeze_dim=1
-        )
+        query_states, key_states = apply_rotary_pos_emb(query_states, key_states, cos, sin, unsqueeze_dim=1)
 
         if past_key_value is not None and past_key_value[0] is not None:
             past_k, past_v = past_key_value
@@ -229,9 +220,7 @@ class Glm4Attention(nn.Layer):
             key_states = paddle.repeat_interleave(key_states, repeat_factor, axis=1)
             value_states = paddle.repeat_interleave(value_states, repeat_factor, axis=1)
 
-        attn_weights = paddle.matmul(
-            query_states, key_states, transpose_y=True
-        ) / (self.head_dim ** 0.5)
+        attn_weights = paddle.matmul(query_states, key_states, transpose_y=True) / (self.head_dim**0.5)
 
         kv_len = key_states.shape[2]
         cur_q_len = query_states.shape[2]
@@ -257,14 +246,10 @@ class Glm4Attention(nn.Layer):
             else:
                 attn_weights = attn_weights + attention_mask
 
-        attn_weights = F.softmax(attn_weights.astype("float32"), axis=-1).astype(
-            query_states.dtype
-        )
+        attn_weights = F.softmax(attn_weights.astype("float32"), axis=-1).astype(query_states.dtype)
         attn_output = paddle.matmul(attn_weights, value_states)
 
-        attn_output = attn_output.transpose([0, 2, 1, 3]).reshape(
-            [bsz, cur_q_len, self.num_heads * self.head_dim]
-        )
+        attn_output = attn_output.transpose([0, 2, 1, 3]).reshape([bsz, cur_q_len, self.num_heads * self.head_dim])
         attn_output = self.o_proj(attn_output)
 
         return attn_output, present
@@ -376,9 +361,7 @@ class Glm4Model(Glm4PretrainedModel):
             config.hidden_size,
             padding_idx=self.padding_idx,
         )
-        self.layers = nn.LayerList(
-            [Glm4DecoderLayer(config) for _ in range(config.num_hidden_layers)]
-        )
+        self.layers = nn.LayerList([Glm4DecoderLayer(config) for _ in range(config.num_hidden_layers)])
         self.norm = Glm4RMSNorm(config.hidden_size, eps=config.rms_norm_eps)
         self.rotary_emb = Glm4RotaryEmbedding(config)
 
@@ -416,9 +399,7 @@ class Glm4Model(Glm4PretrainedModel):
             past_length = past_key_values[0][0].shape[2]
 
         if position_ids is None:
-            position_ids = paddle.arange(
-                past_length, past_length + seq_len, dtype="int64"
-            ).unsqueeze(0).tile([bsz, 1])
+            position_ids = paddle.arange(past_length, past_length + seq_len, dtype="int64").unsqueeze(0).tile([bsz, 1])
 
         position_embeddings = self.rotary_emb(position_ids)
 
@@ -491,9 +472,7 @@ class Glm4ForCausalLM(Glm4PretrainedModel):
                 paddle.zeros_like(position_ids),
             )
         else:
-            position_ids = paddle.arange(
-                input_ids.shape[1], dtype="int64"
-            ).unsqueeze(0).tile([input_ids.shape[0], 1])
+            position_ids = paddle.arange(input_ids.shape[1], dtype="int64").unsqueeze(0).tile([input_ids.shape[0], 1])
 
         if (
             past_key_values is not None
